refactor: change folder structure, issue #16
