+++ backend/src/main/java/com/markethub/features/auth/dto/RegisterRequest.java
@@ -1,7 +1,8 @@
-package com.markethub.dto;
+package com.markethub.features.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,11 +12,16 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class LoginRequest {
+public class RegisterRequest {
+    @NotNull(message = "Name is required")
+    @Size(min = 3, max = 50, message = "Name must be between 3 and 50 characters")
+    private String userName;
+
     @NotNull(message = "Email is required")
     @Email(message = "Email is invalid")
     private String email;
 
     @NotNull(message = "Password is required")
+    @Size(min = 6, message = "Password must be at least 6 characters long")
     private String password;
 }
