+++ backend/src/main/java/com/markethub/features/products/dto/ProductUpdateRequest.java
@@ -1,6 +1,6 @@
-package com.markethub.dto;
+package com.markethub.features.products.dto;
 
-import com.markethub.model.ProductStatus;
+import com.markethub.features.products.models.ProductStatus;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
