+++ backend/src/main/java/com/markethub/features/products/models/ProductStatus.java
@@ -1,4 +1,4 @@
-package com.markethub.model;
+package com.markethub.features.products.models;
 
 public enum ProductStatus {
     ACTIVE,
