+++ backend/src/main/java/com/markethub/features/products/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.markethub.repository;
+package com.markethub.features.products.repository;
 
-import com.markethub.model.Category;
+import com.markethub.features.products.models.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
