+++ backend/src/main/java/com/markethub/features/user/models/Role.java
@@ -1,0 +1,6 @@
+package com.markethub.features.user.models;
+
+public enum Role {
+    USER,
+    ADMIN
+}
