+++ backend/src/main/java/com/markethub/features/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.markethub.repository;
+package com.markethub.features.user.repository;
 
-import com.markethub.model.User;
+import com.markethub.features.user.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 import java.util.UUID;
