--- backend/src/main/java/com/markethub/model/Role.java
@@ -1,6 +1,0 @@
-package com.markethub.model;
-
-public enum Role {
-    USER,
-    ADMIN
-}
